Reindex move out from server module
Moved ClientScrollableHitSource to sub-package to circumvent java
compiler or module system bug (or our abuse of it, not sure).
